Fixed bitmap pixel order issues (Sprite.tone). Fixed Graphics.backcolor. Improved scripts.rb tests.
git-svn-id: https://argss.svn.sourceforge.net/svnroot/argss/trunk@23 a70b67f4-0116-4799-9eb2-a6e6dfe7dbda

+++ src/graphics.h
@@ -67,7 +67,7 @@ namespace Graphics {
 	extern int fps;
 	extern int framerate;
 	extern int framecount;
-	extern unsigned long backcolor;
+	extern Uint32 backcolor;
 	extern int brightness;
 	extern double framerate_interval;
 	extern long creation;
